@@ -6,12 +6,9 @@
     "node": "16"
   },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^13.5.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "jest": {
     "collectCoverageFrom": [
@@ -21,6 +18,7 @@
   "scripts": {
     "cy": "cypress run",
     "cy:open": "cypress open",
+    "test": "vitest",
     "start": "vite",
     "build": "vite build",
     "serve": "vite preview",
@@ -44,6 +42,8 @@
   },
   "devDependencies": {
     "@sheerun/mutationobserver-shim": "0.3.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
     "@vitejs/plugin-react": "^4.0.0",
     "cypress": "^9.6.1",
     "cypress-multi-reporters": "^1.6.0",
@@ -56,6 +56,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "jsdom": "^22.0.0",
     "mocha": "^10.0.0",
     "mochawesome": "^7.1.3",
     "mochawesome-merge": "^4.2.1",
@@ -63,6 +64,7 @@
     "stylelint": "13.8.0",
     "stylelint-order": "4.1.0",
     "vite": "^4.3.5",
-    "vite-plugin-eslint": "^1.8.1"
+    "vite-plugin-eslint": "^1.8.1",
+    "vitest": "^0.31.0"
   }
 }
