@@ -1,7 +1,7 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React from 'react'
+import PropTypes from 'prop-types'
 
-const TableRow = ({ data }) => {
+function TableRow({ data }) {
   const {
     name,
     rotation_period: rotationPeriod,
@@ -16,26 +16,26 @@ const TableRow = ({ data }) => {
     created,
     edited,
     url,
-  } = data;
+  } = data
 
   return (
     <tr>
-      <td data-testid="planet-name">{ name }</td>
-      <td>{ rotationPeriod }</td>
-      <td>{ orbitalPeriod }</td>
-      <td>{ diameter }</td>
-      <td>{ climate }</td>
-      <td>{ gravity }</td>
-      <td>{ terrain }</td>
-      <td>{ surfaceWater }</td>
-      <td>{ population }</td>
-      <td>{ films }</td>
-      <td>{ created }</td>
-      <td>{ edited }</td>
-      <td>{ url }</td>
+      <td data-testid="planet-name">{name}</td>
+      <td>{rotationPeriod}</td>
+      <td>{orbitalPeriod}</td>
+      <td>{diameter}</td>
+      <td>{climate}</td>
+      <td>{gravity}</td>
+      <td>{terrain}</td>
+      <td>{surfaceWater}</td>
+      <td>{population}</td>
+      <td>{films}</td>
+      <td>{created}</td>
+      <td>{edited}</td>
+      <td>{url}</td>
     </tr>
-  );
-};
+  )
+}
 
 TableRow.propTypes = {
   data: PropTypes.shape({
@@ -53,6 +53,6 @@ TableRow.propTypes = {
     edited: PropTypes.string.isRequired,
     url: PropTypes.string.isRequired,
   }).isRequired,
-};
+}
 
-export default TableRow;
+export default TableRow
