@@ -1,1 +1,2 @@
-export const PLANETS_ENDPOINT = "https://swapi.dev/api/planets";
+export const PLANETS_ENDPOINT = 'https://swapi.dev/api/planets'
+export const teste = 'teste'
