@@ -1,27 +1,28 @@
 const applyNumericFilter = (arrayOfObjects, filters) => {
-  const { column, value, operator } = filters;
-  const filteredData = arrayOfObjects.filter((object) => {
-    const objectValue = object[column];
-    const isValueKnown = objectValue !== 'unknown';
+  const { column, value, operator } = filters
+  const filteredData = arrayOfObjects.filter(object => {
+    const objectValue = object[column]
+    const isValueKnown = objectValue !== 'unknown'
     if (isValueKnown) {
-      const parsedObjectValue = parseInt(objectValue, 10);
-      const parsedNumericFilter = parseInt(value, 10);
+      const parsedObjectValue = parseInt(objectValue, 10)
+      const parsedNumericFilter = parseInt(value, 10)
       const parsedFilters = {
         'maior que': parsedObjectValue > parsedNumericFilter,
         'menor que': parsedObjectValue < parsedNumericFilter,
         'igual a': parsedObjectValue === parsedNumericFilter,
-      };
-      const isFilterTrue = parsedFilters[operator];
-      return isFilterTrue;
-    } return false;
-  });
-  return filteredData;
-};
+      }
+      const isFilterTrue = parsedFilters[operator]
+      return isFilterTrue
+    }
+    return false
+  })
+  return filteredData
+}
 
-const filterDataByNumericValues = (dataToBeFiltered, usedFiltersData) => usedFiltersData
-  .reduce((accumulator, currentFilter) => {
-    const result = applyNumericFilter(accumulator, currentFilter);
-    return result;
-  }, dataToBeFiltered);
+const filterDataByNumericValues = (dataToBeFiltered, usedFiltersData) =>
+  usedFiltersData.reduce((accumulator, currentFilter) => {
+    const result = applyNumericFilter(accumulator, currentFilter)
+    return result
+  }, dataToBeFiltered)
 
-export default filterDataByNumericValues;
+export default filterDataByNumericValues
