@@ -1,12 +1,16 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import eslint from 'vite-plugin-eslint';
+/* eslint-disable import/no-extraneous-dependencies */
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+import eslint from 'vite-plugin-eslint'
 
-export default defineConfig(() => {
-  return {
-    build: {
-      outDir: 'build',
-    },
-    plugins: [react(), eslint()],
-  };
-});
+export default defineConfig(() => ({
+  build: {
+    outDir: 'build',
+  },
+  plugins: [react(), eslint()],
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    setupFiles: './src/tests/setup.js',
+  },
+}))
